+++ packages/core-cairo/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0 (2022-06-15)
+
+- Update API format for Cairo. ([#136](https://github.com/OpenZeppelin/contracts-wizard/pull/136))
+- Add functions for getInitialSupply and toUint256. ([#138](https://github.com/OpenZeppelin/contracts-wizard/pull/138))
+
 ## 0.1.0 (2022-05-13)
 
 - Initial API for Cairo. ([#127](https://github.com/OpenZeppelin/contracts-wizard/pull/127))